requirements.txt
streamlit==1.40.0
python-dotenv>=1.2.2
crewai[google-genai]==1.14.3
tavily-python>=0.7.0
langchain-chroma==0.2.6
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
sentence-transformers>=3.0.0